LayoutTests/accessibility/svg-bounds-expected.txt

Test
This test ensures the accessibility bounds of embedded SVG objects are correct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


container location: (607, 107)
Face role: AXRole: AXButton
Face label: AXDescription: face
FaceX: 0
FaceY: 0


Eye role: AXRole: AXButton
Eye label: AXDescription: left-eye
EyeX: 103
EyeY: 148


Nose role: AXRole: AXButton
Nose label: AXDescription: nose
NoseX: 193
NoseY: 206


Mouth role: AXRole: AXButton
Mouth label: AXDescription: smile
MouthX: 115
MouthY: 275


Text role: AXRole: AXStaticText
TextX/10: 15
TextY/10: 11


Image role: AXRole: AXImage
Image label: AXDescription: Test Image
ImageX: 21
ImageY: 21
PASS successfullyParsed is true

TEST COMPLETE
